requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.37
pandas==2.1.4
ta==0.11.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
numpy==1.26.2
finnhub-python==2.4.19